move EmptyResponse type up to top level of api-client
