+++ api-client/src/index.ts
@@ -1,5 +1,5 @@
 // api client entry point
-export type { HostConfig } from './types'
+export type { HostConfig, EmptyResponse } from './types'
 export * from './request'
 export * from './health'
 export * from './sessions'
